Add files via upload
Welcome to the KBC (Kaun Banega Crorepati) simulation project! This repository contains a Python-based implementation of the popular quiz show, KBC. The program emulates the format and style of the show, allowing users to test their knowledge through a series of multiple-choice questions and shows the prize money they won.